[UPDATE] 상대방 정보 저장하는 UserData 함수 제작

+++ Runners_Hi/Runners_Hi/Source/Data/Opponent+CoreDataProperties.swift
@@ -16,10 +16,10 @@ extension Opponent {
         return NSFetchRequest<Opponent>(entityName: "Opponent")
     }
 
-    @NSManaged public var level: Int16
-    @NSManaged public var lose: Int16
+    @NSManaged public var level: Int64
+    @NSManaged public var lose: Int64
     @NSManaged public var nickname: String?
-    @NSManaged public var profileImage: Int16
-    @NSManaged public var win: Int16
+    @NSManaged public var profileImage: Int64
+    @NSManaged public var win: Int64
 
 }
